+++ src/main/resources/com/example/demo/mapper/UserMapper.xml
@@ -25,8 +25,7 @@
 		<if test="birthdate != null">
 			AND birthdate = #{birthdate}
 		</if>
-		ORDER BY
-			id
+		ORDER BY id
 	</select>
 
 	<resultMap id="UserMap" type="com.example.demo.entity.User">
